refactor: split cli_show_interfaces.go per render mode (#4654)
pkg/cli/cli_show_interfaces.go had grown to 1396 LOC, and the RETH /
physical-member display logic repeated across the summary, terse, detail,
and extensive render modes. That duplication is exactly where the #4328
family of fixes kept drifting between modes. Splitting the file per render
mode and pulling the shared RETH/kernel-query helpers into one file removes
the drift surface: the four modes now resolve reth link state, member
attrs, and lease data through a single place.

This is a pure code-motion change. Every top-level declaration was moved
verbatim; no logic, output text, or netlink/sysfs query order changed.

New layout (same package, so unexported helpers stay reachable):

  cli_show_interfaces.go          showInterfaces dispatch + summary,
                                  showInterfacesRethMemberSummary,
                                  showTunnelInterfaces
  cli_show_interfaces_terse.go    showInterfacesTerse
  cli_show_interfaces_detail.go   showInterfacesDetail +
                                  showInterfacesRethDetail (reused by
                                  the extensive view)
  cli_show_interfaces_extensive.go showInterfacesExtensive[Filtered]
  cli_show_interfaces_stats.go    showInterfacesStatistics, showVlans
  cli_show_interfaces_shared.go   dhcpLease, rethMemberLinkState,
                                  rethMemberAttrs, baseIfName

Validation:
  - go/parser decl-diff of origin/master's single file against the union
    of the six new files: identical=14, bodydiff=0, missing=0, extra=0 —
    all 14 top-level declarations byte-for-byte identical.
  - go build ./... clean; go vet ./pkg/cli/... reports only the
    pre-existing cli.go:503 unreachable-code note (untouched file).
  - go test ./pkg/cli/... green, including the #4328 (RETH) and #3654
    (host-inbound) output tests that assert the exact rendered text across
    the summary/detail/extensive/member paths — the visible-output
    byte-identity gate.
  - gofmt clean on all six files.

Closes #4654
