+++ pkg/cli/cli_show_interfaces.go
@@ -3,7 +3,6 @@ package cli
 import (
 	"fmt"
 	"net"
-	"os"
 	"sort"
 	"strconv"
 	"strings"
@@ -12,16 +11,6 @@ import (
 	"github.com/psaab/xpf/pkg/dhcp"
 	"github.com/vishvananda/netlink"
 )
-
-// dhcpLease returns the current DHCP lease for a managed interface, if any.
-// Consumed by `show interfaces` presenters when rendering DHCP-assigned
-// addresses; returns nil when no DHCP client is wired or no lease exists.
-func (c *CLI) dhcpLease(ifaceName string, af dhcp.AddressFamily) *dhcp.Lease {
-	if c.dhcp == nil {
-		return nil
-	}
-	return c.dhcp.LeaseFor(ifaceName, af)
-}
 
 func (c *CLI) showTunnelInterfaces() error {
 	if c.routing == nil {
@@ -466,28 +455,6 @@ func (c *CLI) showInterfaces(args []string) error {
 	return nil
 }
 
-// rethMemberLinkState returns the admin/link ("up"/"down") state for a reth's
-// physical member, best-effort from the kernel. When the device is absent (a
-// peer-owned member, or a test host with no such netdev) admin stays "up" and
-// only link is reported "down" — matching the terse handler (#4328).
-func rethMemberLinkState(member string) (admin, link string) {
-	admin, link = "up", "up"
-	kernelIf := config.LinuxIfName(member)
-	iface, err := net.InterfaceByName(kernelIf)
-	if err != nil {
-		return "up", "down"
-	}
-	if iface.Flags&net.FlagUp == 0 {
-		admin = "down"
-	}
-	if data, err := os.ReadFile("/sys/class/net/" + kernelIf + "/operstate"); err == nil {
-		if strings.TrimSpace(string(data)) != "up" {
-			link = "down"
-		}
-	}
-	return admin, link
-}
-
 // showInterfacesRethMemberSummary renders `show interfaces <member>` for a
 // physical reth member (which is not zoned and so never surfaces through the
 // zone-driven summary walk). It names the reth parent and lists the aggregated
@@ -516,881 +483,4 @@ func (c *CLI) showInterfacesRethMemberSummary(cfg *config.Config, member, reth s
 		fmt.Printf("    aenet --> %s.%d\n", reth, ru.Unit)
 	}
 	fmt.Println()
-}
-
-// dhcpLease returns the DHCP lease for an interface/family, or nil.
-
-func (c *CLI) showInterfacesDetail(filterName string) error {
-	links, err := netlink.LinkList()
-	if err != nil {
-		return fmt.Errorf("listing interfaces: %w", err)
-	}
-
-	sort.Slice(links, func(i, j int) bool {
-		return links[i].Attrs().Name < links[j].Attrs().Name
-	})
-
-	// Build zone + description lookup from active config
-	ifZoneMap := make(map[string]string)
-	ifDescMap := make(map[string]string)
-	cfg := c.store.ActiveConfig()
-	// #4328: reth resolution — a bondless reth has no kernel netdev, so the
-	// netlink walk below never emits it. Build the shared maps so a reth
-	// aggregate is synthesized from config and its physical members are
-	// annotated with their aenet aggregation.
-	var rethMaps config.RethShowMaps
-	if cfg != nil {
-		rethMaps = cfg.RethShowMaps()
-		for _, z := range cfg.Security.Zones {
-			if z == nil { // #3493: tolerant/HA-sync path may carry a nil zone value
-				continue
-			}
-			for _, ifName := range z.Interfaces {
-				ifZoneMap[ifName] = z.Name
-			}
-		}
-		for _, ifc := range cfg.Interfaces.Interfaces {
-			if ifc.Description != "" {
-				ifDescMap[ifc.Name] = ifc.Description
-			}
-		}
-	}
-
-	found := false
-	for _, link := range links {
-		attrs := link.Attrs()
-		if attrs.Name == "lo" {
-			continue
-		}
-		if filterName != "" && attrs.Name != filterName {
-			continue
-		}
-		found = true
-
-		adminUp := attrs.Flags&net.FlagUp != 0
-		operUp := attrs.OperState == netlink.OperUp
-		adminStr := "Disabled"
-		if adminUp {
-			adminStr = "Enabled"
-		}
-		linkStr := "Down"
-		if operUp {
-			linkStr = "Up"
-		}
-		fmt.Printf("Physical interface: %s, %s, Physical link is %s\n", attrs.Name, adminStr, linkStr)
-		if desc, ok := ifDescMap[attrs.Name]; ok {
-			fmt.Printf("  Description: %s\n", desc)
-		}
-		fmt.Printf("  Interface index: %d, SNMP ifIndex: %d\n", attrs.Index, attrs.Index)
-
-		// Link type, MTU, speed, duplex
-		linkType := "Ethernet"
-		if attrs.EncapType != "" {
-			linkType = attrs.EncapType
-		}
-		speedStr := ""
-		if speed := readLinkSpeed(attrs.Name); speed > 0 {
-			speedStr = ", Speed: " + formatSpeed(speed)
-		}
-		duplexStr := ""
-		if d := readLinkDuplex(attrs.Name); d != "" {
-			duplexStr = ", Duplex: " + formatDuplex(d)
-		}
-		fmt.Printf("  Link-level type: %s, MTU: %d%s%s\n", linkType, attrs.MTU, speedStr, duplexStr)
-
-		if len(attrs.HardwareAddr) > 0 {
-			fmt.Printf("  Current address: %s\n", attrs.HardwareAddr)
-		}
-		if zone, ok := ifZoneMap[attrs.Name]; ok {
-			fmt.Printf("  Security zone: %s\n", zone)
-		}
-
-		// Logical interface with flags and addresses
-		var flags []string
-		if adminUp {
-			flags = append(flags, "Up")
-		}
-		if attrs.RawFlags&0x2 != 0 { // IFF_BROADCAST
-			flags = append(flags, "BROADCAST")
-		}
-		if attrs.OperState == netlink.OperUp {
-			flags = append(flags, "RUNNING")
-		}
-		if attrs.RawFlags&0x1000 != 0 { // IFF_MULTICAST
-			flags = append(flags, "MULTICAST")
-		}
-		fmt.Printf("  Logical interface %s.0\n", attrs.Name)
-		if len(flags) > 0 {
-			fmt.Printf("    Flags: %s\n", strings.Join(flags, " "))
-		}
-
-		// #4328: a physical reth member carries the reth's logical units as
-		// aenet aggregation — surface that membership here (terse already does).
-		if rethName, ok := rethMaps.LookupMember(attrs.Name); ok && cfg != nil {
-			fmt.Printf("    Redundant-ethernet: member of %s\n", rethName)
-			for _, ru := range cfg.RethShowUnits(rethName) {
-				fmt.Printf("    Logical interface %s.%d --> aenet %s.%d\n",
-					attrs.Name, ru.Unit, rethName, ru.Unit)
-			}
-		}
-
-		addrs, _ := netlink.AddrList(link, netlink.FAMILY_ALL)
-		if len(addrs) > 0 {
-			fmt.Println("    Addresses:")
-			for _, a := range addrs {
-				fmt.Printf("      %s\n", a.IPNet)
-			}
-		}
-
-		// Traffic statistics
-		if s := attrs.Statistics; s != nil {
-			fmt.Println("  Traffic statistics:")
-			fmt.Printf("    Input  packets:             %12d\n", s.RxPackets)
-			fmt.Printf("    Output packets:             %12d\n", s.TxPackets)
-			fmt.Printf("    Input  bytes:               %12d\n", s.RxBytes)
-			fmt.Printf("    Output bytes:               %12d\n", s.TxBytes)
-			fmt.Printf("    Input  errors:              %12d\n", s.RxErrors)
-			fmt.Printf("    Output errors:              %12d\n", s.TxErrors)
-		}
-		fmt.Println()
-	}
-
-	// #4328: render bondless reth aggregates (no kernel netdev) and, when the
-	// filter names a reth member whose kernel device is absent, a synthetic
-	// member block. This runs after the netlink walk so an unfiltered
-	// `show interfaces detail` lists reths alongside the physical members.
-	if cfg != nil {
-		if c.showInterfacesRethDetail(cfg, rethMaps, filterName, found) {
-			found = true
-		}
-	}
-
-	if filterName != "" && !found {
-		fmt.Printf("Interface %s not found\n", filterName)
-	}
-	return nil
-}
-
-// baseIfName strips a dotted unit suffix ("reth0.50" -> "reth0").
-func baseIfName(name string) string {
-	if i := strings.IndexByte(name, '.'); i >= 0 {
-		return name[:i]
-	}
-	return name
-}
-
-// rethMemberAttrs returns the netlink attributes of a reth's physical member,
-// best-effort. ok is false when the member device is absent (peer-owned or a
-// test host) so callers render from config alone (#4328).
-func rethMemberAttrs(member string) (attrs *netlink.LinkAttrs, ok bool) {
-	link, err := netlink.LinkByName(config.LinuxIfName(member))
-	if err != nil {
-		return nil, false
-	}
-	return link.Attrs(), true
-}
-
-// showInterfacesRethDetail renders the `show interfaces ... detail` view for
-// bondless reth aggregates (which have no kernel netdev) and, when the filter
-// names a reth member whose kernel device is absent, a synthetic member block.
-// alreadyFound reports whether the netlink walk already rendered the filtered
-// interface (a real member device). Returns true when anything was rendered.
-func (c *CLI) showInterfacesRethDetail(cfg *config.Config, maps config.RethShowMaps, filterName string, alreadyFound bool) bool {
-	// Zone lookup keyed by logical ref (reth0.50 -> zone).
-	ifZone := make(map[string]string)
-	for _, z := range cfg.Security.Zones {
-		if z == nil {
-			continue
-		}
-		for _, ifName := range z.Interfaces {
-			ifZone[ifName] = z.Name
-		}
-	}
-
-	rendered := false
-	reths := make([]string, 0, len(maps.RethToPhys))
-	for reth := range maps.RethToPhys {
-		reths = append(reths, reth)
-	}
-	sort.Strings(reths)
-	for _, reth := range reths {
-		if filterName != "" && baseIfName(filterName) != reth {
-			continue
-		}
-		member := maps.RethToPhys[reth]
-		attrs, haveDev := rethMemberAttrs(member)
-		adminStr, linkStr := "Disabled", "Down"
-		mtu := 0
-		var hwAddr net.HardwareAddr
-		if haveDev {
-			if attrs.Flags&net.FlagUp != 0 {
-				adminStr = "Enabled"
-			}
-			if attrs.OperState == netlink.OperUp {
-				linkStr = "Up"
-			}
-			mtu = attrs.MTU
-			hwAddr = attrs.HardwareAddr
-		}
-		fmt.Printf("Physical interface: %s, %s, Physical link is %s\n", reth, adminStr, linkStr)
-		if ifc, ok := cfg.Interfaces.Interfaces[reth]; ok && ifc.Description != "" {
-			fmt.Printf("  Description: %s\n", ifc.Description)
-		}
-		fmt.Printf("  Redundant-ethernet: aggregate over member %s\n", member)
-		fmt.Printf("  Link-level type: Ethernet, MTU: %d\n", mtu)
-		if len(hwAddr) > 0 {
-			fmt.Printf("  Current address: %s\n", hwAddr)
-		}
-		for _, ru := range cfg.RethShowUnits(reth) {
-			fmt.Printf("  Logical interface %s.%d", reth, ru.Unit)
-			if ru.VlanID > 0 {
-				fmt.Printf(" VLAN-Tag [ 0x8100.%d ]", ru.VlanID)
-			}
-			fmt.Println()
-			if zone, ok := ifZone[fmt.Sprintf("%s.%d", reth, ru.Unit)]; ok {
-				fmt.Printf("    Security zone: %s\n", zone)
-			}
-			if len(ru.V4Addrs) > 0 || len(ru.V6Addrs) > 0 {
-				fmt.Println("    Addresses:")
-				for _, a := range ru.V4Addrs {
-					fmt.Printf("      %s\n", a)
-				}
-				for _, a := range ru.V6Addrs {
-					fmt.Printf("      %s\n", a)
-				}
-			}
-		}
-		fmt.Println()
-		rendered = true
-	}
-
-	// A reth member whose kernel device is absent never surfaces through the
-	// netlink walk; render its aenet membership synthetically so the filtered
-	// lookup is not empty.
-	if !alreadyFound && !rendered && filterName != "" {
-		if reth, ok := maps.LookupMember(filterName); ok {
-			member := baseIfName(filterName)
-			admin, link := rethMemberLinkState(member)
-			adminStr, linkStr := "Enabled", "Up"
-			if admin == "down" {
-				adminStr = "Disabled"
-			}
-			if link == "down" {
-				linkStr = "Down"
-			}
-			fmt.Printf("Physical interface: %s, %s, Physical link is %s\n", member, adminStr, linkStr)
-			if ifc, ok := cfg.Interfaces.Interfaces[member]; ok && ifc.Description != "" {
-				fmt.Printf("  Description: %s\n", ifc.Description)
-			}
-			fmt.Printf("  Redundant-ethernet: member of %s\n", reth)
-			for _, ru := range cfg.RethShowUnits(reth) {
-				fmt.Printf("  Logical interface %s.%d --> aenet %s.%d\n", member, ru.Unit, reth, ru.Unit)
-			}
-			fmt.Println()
-			rendered = true
-		}
-	}
-	return rendered
-}
-
-func (c *CLI) showInterfacesTerse() error {
-	cfg := c.store.ActiveConfig()
-	if cfg == nil {
-		fmt.Println("no active configuration")
-		return nil
-	}
-
-	// Build RETH mappings (shared resolver, #4328 — the same maps back the
-	// summary/detail/extensive paths so they can never drift from terse again).
-	rethMaps := cfg.RethShowMaps()
-	physToReth := rethMaps.PhysToReth // physical member → reth parent
-	rethToPhys := rethMaps.RethToPhys // reth → physical member
-
-	type ifUnit struct {
-		physName string
-		unitNum  int
-		vlanID   int
-	}
-	var units []ifUnit
-
-	for physName, ifCfg := range cfg.Interfaces.Interfaces {
-		if rethName, ok := physToReth[physName]; ok {
-			// Physical RETH member: inherit units from RETH parent
-			if rethCfg, ok := cfg.Interfaces.Interfaces[rethName]; ok {
-				for unitNum, unit := range rethCfg.Units {
-					units = append(units, ifUnit{physName: physName, unitNum: unitNum, vlanID: unit.VlanID})
-				}
-			}
-		} else {
-			for unitNum, unit := range ifCfg.Units {
-				units = append(units, ifUnit{physName: physName, unitNum: unitNum, vlanID: unit.VlanID})
-			}
-		}
-	}
-
-	// Add peer node interfaces (cluster mode).
-	peerIfaces := make(map[string]bool)
-	peerLinkUp := make(map[string]bool)
-	if c.cluster != nil {
-		peerNodeID := -1
-		if c.cluster.PeerAlive() {
-			peerNodeID = c.cluster.PeerNodeID()
-		} else if cfg.Chassis.Cluster != nil {
-			localID := c.cluster.NodeID()
-			for _, rg := range cfg.Chassis.Cluster.RedundancyGroups {
-				for nid := range rg.NodePriorities {
-					if nid != localID {
-						peerNodeID = nid
-						break
-					}
-				}
-				if peerNodeID >= 0 {
-					break
-				}
-			}
-		}
-		if peerNodeID >= 0 {
-			if peerMons := c.cluster.PeerMonitorStatuses(); peerMons != nil {
-				for _, pm := range peerMons {
-					peerLinkUp[pm.Interface] = pm.Up
-				}
-			}
-			tree := c.store.ActiveTree()
-			if tree != nil {
-				// Lenient compile — the active tree was already tolerated
-				// on load (Store.Load compiles with the tolerant-path
-				// downgrades, e.g. #1798 control-char sanitize). A strict
-				// re-compile here could error on such a legacy config and
-				// silently drop peer-interface display via the err==nil
-				// guard below.
-				peerCfg, err := config.CompileConfigForNodeLenient(tree, peerNodeID)
-				if err == nil {
-					for physName, ifCfg := range peerCfg.Interfaces.Interfaces {
-						if _, isLocal := cfg.Interfaces.Interfaces[physName]; isLocal {
-							continue
-						}
-						peerIfaces[physName] = true
-						if ifCfg.RedundantParent != "" {
-							physToReth[physName] = ifCfg.RedundantParent
-							if rethCfg, ok := peerCfg.Interfaces.Interfaces[ifCfg.RedundantParent]; ok {
-								for unitNum, unit := range rethCfg.Units {
-									units = append(units, ifUnit{physName: physName, unitNum: unitNum, vlanID: unit.VlanID})
-								}
-							}
-						} else {
-							for unitNum, unit := range ifCfg.Units {
-								units = append(units, ifUnit{physName: physName, unitNum: unitNum, vlanID: unit.VlanID})
-							}
-						}
-					}
-				}
-			}
-		}
-	}
-
-	sort.Slice(units, func(i, j int) bool {
-		if units[i].physName != units[j].physName {
-			return units[i].physName < units[j].physName
-		}
-		return units[i].unitNum < units[j].unitNum
-	})
-
-	fmt.Printf("%-24s%-6s%-6s%-9s%-22s\n", "Interface", "Admin", "Link", "Proto", "Local")
-
-	printedPhys := make(map[string]bool)
-
-	for _, u := range units {
-		isPeer := peerIfaces[u.physName]
-
-		if !printedPhys[u.physName] {
-			printedPhys[u.physName] = true
-			admin := "up"
-			link := "up"
-			if isPeer {
-				// Peer interface: use heartbeat monitor data.
-				if up, ok := peerLinkUp[u.physName]; ok {
-					if !up {
-						link = "down"
-					}
-				} else if c.cluster != nil && !c.cluster.PeerAlive() {
-					link = "down"
-				}
-			} else {
-				// Local interface: query kernel.
-				// Check config-level disable flag
-				if ifCfg, ok := cfg.Interfaces.Interfaces[u.physName]; ok && ifCfg.Disable {
-					admin = "down"
-				}
-				// For RETH interfaces, get status from physical member
-				statusIf := u.physName
-				if phys, ok := rethToPhys[u.physName]; ok {
-					statusIf = phys
-				}
-				kernelIf := config.LinuxIfName(statusIf)
-				iface, err := net.InterfaceByName(kernelIf)
-				if err != nil {
-					link = "down"
-				} else {
-					if iface.Flags&net.FlagUp == 0 {
-						if admin == "up" {
-							admin = "down" // kernel says down
-						}
-					}
-					data, err := os.ReadFile("/sys/class/net/" + kernelIf + "/operstate")
-					if err == nil && strings.TrimSpace(string(data)) != "up" {
-						link = "down"
-					}
-				}
-			}
-			fmt.Printf("%-24s%-6s%-6s\n", u.physName, admin, link)
-		}
-
-		logicalName := fmt.Sprintf("%s.%d", u.physName, u.unitNum)
-
-		// Physical RETH member: show aenet --> rethN.M
-		if rethName, ok := physToReth[u.physName]; ok {
-			admin := "up"
-			link := "up"
-			if isPeer {
-				if up, ok := peerLinkUp[u.physName]; ok {
-					if !up {
-						link = "down"
-					}
-				} else if c.cluster != nil && !c.cluster.PeerAlive() {
-					link = "down"
-				}
-			} else {
-				kernelIf := config.LinuxIfName(u.physName)
-				iface, err := net.InterfaceByName(kernelIf)
-				if err != nil {
-					link = "down"
-				} else {
-					if iface.Flags&net.FlagUp == 0 {
-						admin = "down"
-					}
-					data, err := os.ReadFile("/sys/class/net/" + kernelIf + "/operstate")
-					if err == nil && strings.TrimSpace(string(data)) != "up" {
-						link = "down"
-					}
-				}
-			}
-			rethLogical := fmt.Sprintf("%s.%d", rethName, u.unitNum)
-			fmt.Printf("%-24s%-6s%-6s%-9s%s\n", logicalName, admin, link, "aenet", "--> "+rethLogical)
-			continue
-		}
-
-		// RETH interface: get addresses from config, status from physical member
-		if physMember, ok := rethToPhys[u.physName]; ok {
-			var v4Addrs, v6Addrs []string
-			if ifCfg, ok := cfg.Interfaces.Interfaces[u.physName]; ok {
-				if unit, ok := ifCfg.Units[u.unitNum]; ok {
-					for _, addr := range unit.Addresses {
-						ip, _, err := net.ParseCIDR(addr)
-						if err != nil {
-							continue
-						}
-						if ip.To4() != nil {
-							v4Addrs = append(v4Addrs, addr)
-						} else {
-							v6Addrs = append(v6Addrs, addr)
-						}
-					}
-				}
-			}
-			admin := "up"
-			link := "up"
-			kernelPhys := config.LinuxIfName(physMember)
-			iface, err := net.InterfaceByName(kernelPhys)
-			if err != nil {
-				link = "down"
-			} else {
-				if iface.Flags&net.FlagUp == 0 {
-					admin = "down"
-				}
-				data, err := os.ReadFile("/sys/class/net/" + kernelPhys + "/operstate")
-				if err == nil && strings.TrimSpace(string(data)) != "up" {
-					link = "down"
-				}
-			}
-			firstProto := ""
-			firstAddr := ""
-			if len(v4Addrs) > 0 {
-				firstProto = "inet"
-				firstAddr = v4Addrs[0]
-			} else if len(v6Addrs) > 0 {
-				firstProto = "inet6"
-				firstAddr = v6Addrs[0]
-			}
-			fmt.Printf("%-24s%-6s%-6s%-9s%-22s\n", logicalName, admin, link, firstProto, firstAddr)
-			for i := 1; i < len(v4Addrs); i++ {
-				fmt.Printf("%-36s%-9s%-22s\n", "", "inet", v4Addrs[i])
-			}
-			startIdx := 0
-			if firstProto == "inet6" {
-				startIdx = 1
-			}
-			for i := startIdx; i < len(v6Addrs); i++ {
-				fmt.Printf("%-36s%-9s%-22s\n", "", "inet6", v6Addrs[i])
-			}
-			continue
-		}
-
-		// Normal interface: get addresses from kernel
-		lookupName := config.LinuxIfName(u.physName)
-		if u.vlanID > 0 {
-			lookupName = fmt.Sprintf("%s.%d", config.LinuxIfName(u.physName), u.vlanID)
-		}
-
-		var v4Addrs, v6Addrs []string
-		liface, err := net.InterfaceByName(lookupName)
-		if err != nil {
-			liface, err = net.InterfaceByName(config.LinuxIfName(u.physName))
-		}
-		if err == nil {
-			addrs, _ := liface.Addrs()
-			for _, addr := range addrs {
-				ipNet, ok := addr.(*net.IPNet)
-				if !ok {
-					continue
-				}
-				ones, _ := ipNet.Mask.Size()
-				addrStr := fmt.Sprintf("%s/%d", ipNet.IP, ones)
-				if ipNet.IP.To4() != nil {
-					v4Addrs = append(v4Addrs, addrStr)
-				} else {
-					v6Addrs = append(v6Addrs, addrStr)
-				}
-			}
-		}
-
-		admin := "up"
-		link := "up"
-		if liface == nil {
-			link = "down"
-		} else if liface.Flags&net.FlagUp == 0 {
-			admin = "down"
-		}
-
-		firstProto := ""
-		firstAddr := ""
-		if len(v4Addrs) > 0 {
-			firstProto = "inet"
-			firstAddr = v4Addrs[0]
-		} else if len(v6Addrs) > 0 {
-			firstProto = "inet6"
-			firstAddr = v6Addrs[0]
-		}
-
-		fmt.Printf("%-24s%-6s%-6s%-9s%-22s\n", logicalName, admin, link, firstProto, firstAddr)
-
-		for i := 1; i < len(v4Addrs); i++ {
-			fmt.Printf("%-36s%-9s%-22s\n", "", "inet", v4Addrs[i])
-		}
-		startIdx := 0
-		if firstProto == "inet6" {
-			startIdx = 1
-		}
-		for i := startIdx; i < len(v6Addrs); i++ {
-			fmt.Printf("%-36s%-9s%-22s\n", "", "inet6", v6Addrs[i])
-		}
-	}
-
-	return nil
-}
-
-// showInterfacesExtensive shows detailed per-interface statistics including
-// all error counters, queue depths, and ethtool-style information.
-
-func (c *CLI) showInterfacesExtensive() error {
-	return c.showInterfacesExtensiveFiltered("")
-}
-
-func (c *CLI) showInterfacesExtensiveFiltered(filterName string) error {
-	links, err := netlink.LinkList()
-	if err != nil {
-		return fmt.Errorf("listing interfaces: %w", err)
-	}
-
-	sort.Slice(links, func(i, j int) bool {
-		return links[i].Attrs().Name < links[j].Attrs().Name
-	})
-
-	// Build zone lookup from active config
-	ifZoneMap := make(map[string]string)
-	ifDescMap := make(map[string]string)
-	ifCfgMap := make(map[string]*config.InterfaceConfig)
-	cfg := c.store.ActiveConfig()
-	// #4328: reth resolution — synthesize bondless reth aggregates (no kernel
-	// netdev) and annotate physical members with their aenet aggregation.
-	var rethMaps config.RethShowMaps
-	if cfg != nil {
-		rethMaps = cfg.RethShowMaps()
-		for _, z := range cfg.Security.Zones {
-			if z == nil { // #3493: tolerant/HA-sync path may carry a nil zone value
-				continue
-			}
-			for _, ifName := range z.Interfaces {
-				ifZoneMap[ifName] = z.Name
-			}
-		}
-		for _, ifc := range cfg.Interfaces.Interfaces {
-			ifCfgMap[ifc.Name] = ifc
-			if ifc.Description != "" {
-				ifDescMap[ifc.Name] = ifc.Description
-			}
-		}
-	}
-
-	found := false
-	for _, link := range links {
-		attrs := link.Attrs()
-		if attrs.Name == "lo" {
-			continue
-		}
-		if filterName != "" && attrs.Name != filterName {
-			continue
-		}
-		found = true
-
-		// State
-		adminUp := attrs.Flags&net.FlagUp != 0
-		operUp := attrs.OperState == netlink.OperUp
-		adminStr := "Disabled"
-		if adminUp {
-			adminStr = "Enabled"
-		}
-		linkStr := "Down"
-		if operUp {
-			linkStr = "Up"
-		}
-		fmt.Printf("Physical interface: %s, %s, Physical link is %s\n", attrs.Name, adminStr, linkStr)
-		if desc, ok := ifDescMap[attrs.Name]; ok {
-			fmt.Printf("  Description: %s\n", desc)
-		}
-		if zone, ok := ifZoneMap[attrs.Name]; ok {
-			fmt.Printf("  Security zone: %s\n", zone)
-		}
-		// #4328: annotate a physical reth member with its aenet aggregation.
-		if rethName, ok := rethMaps.LookupMember(attrs.Name); ok && cfg != nil {
-			fmt.Printf("  Redundant-ethernet: member of %s\n", rethName)
-			for _, ru := range cfg.RethShowUnits(rethName) {
-				fmt.Printf("  Logical interface %s.%d --> aenet %s.%d\n",
-					attrs.Name, ru.Unit, rethName, ru.Unit)
-			}
-		}
-		if ifCfg, ok := ifCfgMap[attrs.Name]; ok {
-			if ifCfg.Speed != "" {
-				fmt.Printf("  Configured speed: %s\n", ifCfg.Speed)
-			}
-			if ifCfg.Duplex != "" {
-				fmt.Printf("  Configured duplex: %s\n", ifCfg.Duplex)
-			}
-		}
-
-		// Type + speed + MTU
-		linkType := "Ethernet"
-		if attrs.EncapType != "" {
-			linkType = attrs.EncapType
-		}
-		var linkExtras []string
-		if speed := readLinkSpeed(attrs.Name); speed > 0 {
-			linkExtras = append(linkExtras, "Speed: "+formatSpeed(speed))
-		}
-		duplexStr := "Full-duplex"
-		if duplex := readLinkDuplex(attrs.Name); duplex != "" {
-			duplexStr = formatDuplex(duplex)
-		}
-		linkExtras = append(linkExtras, "Link-mode: "+duplexStr)
-		fmt.Printf("  Link-level type: %s, MTU: %d, %s\n",
-			linkType, attrs.MTU, strings.Join(linkExtras, ", "))
-
-		// MAC
-		if len(attrs.HardwareAddr) > 0 {
-			fmt.Printf("  Current address: %s, Hardware address: %s\n",
-				attrs.HardwareAddr, attrs.HardwareAddr)
-		}
-
-		// Device flags
-		var flags []string
-		flags = append(flags, "Present")
-		if operUp {
-			flags = append(flags, "Running")
-		}
-		if !adminUp {
-			flags = append(flags, "Down")
-		}
-		fmt.Printf("  Device flags   : %s\n", strings.Join(flags, " "))
-		fmt.Printf("  Interface index: %d, SNMP ifIndex: %d\n", attrs.Index, attrs.Index)
-
-		if attrs.TxQLen > 0 {
-			fmt.Printf("  Link type      : %s, TxQueueLen: %d\n", attrs.EncapType, attrs.TxQLen)
-		}
-
-		// Detailed statistics
-		if s := attrs.Statistics; s != nil {
-			fmt.Println("  Traffic statistics:")
-			fmt.Printf("    Input:  %d bytes, %d packets\n", s.RxBytes, s.RxPackets)
-			fmt.Printf("    Output: %d bytes, %d packets\n", s.TxBytes, s.TxPackets)
-			fmt.Println("  Input errors:")
-			fmt.Printf("    Errors: %d, Drops: %d, Overruns: %d, Frame: %d\n",
-				s.RxErrors, s.RxDropped, s.RxOverErrors, s.RxFrameErrors)
-			fmt.Printf("    FIFO errors: %d, Missed: %d, Compressed: %d\n",
-				s.RxFifoErrors, s.RxMissedErrors, s.RxCompressed)
-			fmt.Println("  Output errors:")
-			fmt.Printf("    Errors: %d, Drops: %d, Carrier: %d, Collisions: %d\n",
-				s.TxErrors, s.TxDropped, s.TxCarrierErrors, s.Collisions)
-			fmt.Printf("    FIFO errors: %d, Heartbeat: %d, Compressed: %d\n",
-				s.TxFifoErrors, s.TxHeartbeatErrors, s.TxCompressed)
-			if s.Multicast > 0 {
-				fmt.Printf("    Multicast: %d\n", s.Multicast)
-			}
-		}
-
-		// BPF traffic counters (XDP/TC level)
-		if c.dp != nil && c.dp.IsLoaded() {
-			if ctrs, err := c.dp.ReadInterfaceCounters(attrs.Index); err == nil && (ctrs.RxPackets > 0 || ctrs.TxPackets > 0) {
-				fmt.Println("  BPF statistics:")
-				fmt.Printf("    Input:  %d packets, %d bytes\n", ctrs.RxPackets, ctrs.RxBytes)
-				fmt.Printf("    Output: %d packets, %d bytes\n", ctrs.TxPackets, ctrs.TxBytes)
-			}
-		}
-
-		// Addresses
-		addrs, _ := netlink.AddrList(link, netlink.FAMILY_ALL)
-		if len(addrs) > 0 {
-			var v4, v6 []string
-			for _, a := range addrs {
-				if a.IP.To4() != nil {
-					v4 = append(v4, a.IPNet.String())
-				} else {
-					v6 = append(v6, a.IPNet.String())
-				}
-			}
-			if len(v4) > 0 {
-				fmt.Printf("  Protocol inet, MTU: %d\n", attrs.MTU)
-				for _, a := range v4 {
-					fmt.Printf("    Local: %s\n", a)
-				}
-			}
-			if len(v6) > 0 {
-				fmt.Printf("  Protocol inet6, MTU: %d\n", attrs.MTU)
-				for _, a := range v6 {
-					flags := "Is-Preferred Is-Primary"
-					if strings.HasPrefix(a, "fe80:") {
-						flags = "Is-Preferred"
-					}
-					fmt.Printf("    Local: %s, Flags: %s\n", a, flags)
-				}
-			}
-		}
-		fmt.Println()
-	}
-
-	// #4328: bondless reth aggregates + absent-device members (see the detail
-	// path). The extensive view reuses the same synthesized block.
-	if cfg != nil {
-		c.showInterfacesRethDetail(cfg, rethMaps, filterName, found)
-	}
-	return nil
-}
-
-func (c *CLI) showInterfacesStatistics() error {
-	links, err := netlink.LinkList()
-	if err != nil {
-		return fmt.Errorf("listing links: %w", err)
-	}
-
-	sort.Slice(links, func(i, j int) bool {
-		return links[i].Attrs().Name < links[j].Attrs().Name
-	})
-
-	fmt.Printf("%-16s %15s %15s %15s %15s %10s %10s\n",
-		"Interface", "Input packets", "Input bytes", "Output packets", "Output bytes", "In errors", "Out errors")
-
-	for _, l := range links {
-		name := l.Attrs().Name
-		if name == "lo" || strings.HasPrefix(name, "vrf-") ||
-			strings.HasPrefix(name, "xfrm") || strings.HasPrefix(name, "gre-") {
-			continue
-		}
-		stats := l.Attrs().Statistics
-		if stats == nil {
-			continue
-		}
-		fmt.Printf("%-16s %15d %15d %15d %15d %10d %10d\n",
-			name, stats.RxPackets, stats.RxBytes, stats.TxPackets, stats.TxBytes,
-			stats.RxErrors, stats.TxErrors)
-	}
-	return nil
-}
-
-// showVlans displays VLAN assignments per interface (like Junos "show vlans").
-
-func (c *CLI) showVlans() error {
-	cfg := c.store.ActiveConfig()
-	if cfg == nil {
-		fmt.Println("No active configuration")
-		return nil
-	}
-
-	// Build zone lookup: interface name → zone name
-	ifZone := make(map[string]string)
-	for zoneName, zone := range cfg.Security.Zones {
-		if zone == nil { // #3493: tolerant/HA-sync path may carry a nil zone value
-			continue
-		}
-		for _, iface := range zone.Interfaces {
-			ifZone[iface] = zoneName
-		}
-	}
-
-	// Collect VLAN entries
-	type vlanEntry struct {
-		iface  string
-		unit   int
-		vlanID int
-		zone   string
-		trunk  bool
-	}
-	var entries []vlanEntry
-	for _, ifc := range cfg.Interfaces.Interfaces {
-		for unitNum, unit := range ifc.Units {
-			if unit.VlanID > 0 || ifc.VlanTagging {
-				zone := ifZone[ifc.Name]
-				entries = append(entries, vlanEntry{
-					iface:  ifc.Name,
-					unit:   unitNum,
-					vlanID: unit.VlanID,
-					zone:   zone,
-					trunk:  ifc.VlanTagging,
-				})
-			}
-		}
-	}
-
-	if len(entries) == 0 {
-		fmt.Println("No VLANs configured")
-		return nil
-	}
-
-	sort.Slice(entries, func(i, j int) bool {
-		if entries[i].iface != entries[j].iface {
-			return entries[i].iface < entries[j].iface
-		}
-		return entries[i].unit < entries[j].unit
-	})
-
-	fmt.Printf("%-16s %-6s %-8s %-12s %s\n", "Interface", "Unit", "VLAN ID", "Zone", "Mode")
-	for _, e := range entries {
-		mode := "access"
-		if e.trunk {
-			mode = "trunk"
-		}
-		vid := fmt.Sprintf("%d", e.vlanID)
-		if e.vlanID == 0 {
-			vid = "native"
-		}
-		fmt.Printf("%-16s %-6d %-8s %-12s %s\n", e.iface, e.unit, vid, e.zone, mode)
-	}
-	return nil
 }
